Moved all business logic to container components
