+++ src/components/attributes/AttributesContainer.js
@@ -2,7 +2,17 @@ import React, { PureComponent } from 'react'
 import Attributes from './Attributes'
 
 export class AttributesContainer extends PureComponent {
+  constructor() {
+    super()
 
+    this.getAttrType = this.getAttrType.bind(this)
+  }
+
+  getAttrType( attrType ,attributesValueSwatch, attributesValueText ) {
+    const currAttrType = attrType === "swatch" ? attributesValueSwatch : attributesValueText
+
+    return currAttrType
+  }
   render() {
     const {
         attrName,
@@ -19,6 +29,7 @@ export class AttributesContainer extends PureComponent {
         index={index}
         attributesValueSwatch={attributesValueSwatch}
         attributesValueText={attributesValueText}
+        getAttrType={this.getAttrType}
       />
     )
   }
