+++ src/components/body-section/BodySectionContainer.js
@@ -1,27 +1,76 @@
-import React from "react";
-import { PureComponent } from "react";
-import BodySection from "./BodySection";
+import React, { PureComponent } from 'react'
+import BodySection from './BodySection'
+import ProductContainer from "../product/ProductContainer";
+import { Query } from "react-apollo";
 
-export default class BodySectionContainer extends PureComponent {
+export class BodySectionContainer extends PureComponent {
+    constructor() {
+        super();
+    
+        this.productQuery = this.productQuery.bind(this);
+        this.printProducts = this.printProducts.bind(this);
+        this.capitalizeFirstLetter = this.capitalizeFirstLetter.bind(this);
+      }
+    
+      // The product query...
+      productQuery(PRODUCT_QUERY, currencyIdx, cartItemsHandler, currencySymbol) {
+        const productQuery = (
+          <Query query={PRODUCT_QUERY}>
+            {({ loading, data }) => {
+              if (loading) return null;
+              const allProducts = this.printProducts(
+                data,
+                currencyIdx,
+                cartItemsHandler,
+                currencySymbol
+              );
+    
+              return allProducts;
+            }}
+          </Query>
+        );
+    
+        return productQuery;
+      }
+    
+      // Returns the product component...
+      printProducts(data, currencyIdx, cartItemsHandler, currencySymbol) {
+        const allProducts = data.category.products.map((product, idx) => {
+          return (
+            <ProductContainer
+              key={idx}
+              product={product}
+              currencyIdx={currencyIdx}
+              cartItemsHandler={cartItemsHandler}
+              currencySymbol={currencySymbol}
+            />
+          );
+        });
+    
+        return allProducts;
+      }
+    
+      // Capitalizes the first letter of the category title...
+      capitalizeFirstLetter(category) {
+        return category.charAt(0).toUpperCase() + category.slice(1);
+      }
+
   render() {
-    const {
-      match,
-      category,
-      currencyIdx,
-      cartItemsHandler,
-      cartCountHandler,
-      currencySymbol,
-    } = this.props;
+    const { category, currencyIdx, cartItemsHandler, currencySymbol, match } =
+      this.props;
 
     return (
       <BodySection
-        category={ category }
         currencyIdx={ currencyIdx }
         cartItemsHandler={ cartItemsHandler }
-        cartCountHandler={ cartCountHandler }
         currencySymbol={ currencySymbol }
+        category={ category }
         match={ match }
+        productQuery={this.productQuery}
+        capitalizeFirstLetter={this.capitalizeFirstLetter}
       />
-    );
+    )
   }
 }
+
+export default BodySectionContainer
