+++ src/components/clickable-attribute/ClickableAttributeContainer.js
@@ -2,6 +2,15 @@ import React, { PureComponent } from "react";
 import ClickableAttribute from "./ClickableAttribute";
 
 export class ClickableAttributeContainer extends PureComponent {
+  constructor() {
+    super();
+
+    this.attributeContent = this.attributeContent.bind(this);
+  }
+  attributeContent(attrType, swatch, text) {
+    return attrType === "swatch" ? swatch : text;
+  }
+
   render() {
     const {
         attrName,
@@ -18,6 +27,7 @@ export class ClickableAttributeContainer extends PureComponent {
         index={index}
         attributesValueSwatch={attributesValueSwatch}
         attributesValueText={attributesValueText}
+        attributeContent={this.attributeContent}
       />
     );
   }
