+++ src/pages/homepage/HomepageContainer.js
@@ -2,6 +2,21 @@ import React, { PureComponent } from 'react'
 import Homepage from './Homepage'
 
 export class HomepageContainer extends PureComponent {
+  constructor( props ) {
+    super( props );
+    this.state = {
+      currencyIndex: 0,
+    };
+
+    this.updateCurrencyHandler = this.updateCurrencyHandler.bind( this );
+  }
+
+  // Updating the currency...
+  updateCurrencyHandler( e ) {
+    const idx = Number( e.currentTarget.dataset.currindex );
+    this.setState({ currencyIndex: idx });
+  }
+
   render() {
     const {
         cartItems,
@@ -37,6 +52,8 @@ export class HomepageContainer extends PureComponent {
         category={ category }
         getTotalHandler={ getTotalHandler }
         setTotalHandler={ setTotalHandler }
+        updateCurrencyHandler={this.updateCurrencyHandler}
+        currencyIndex={this.state.currencyIndex}
         match={match}
       />
     )
