+++ src/pages/productView/ProductViewContainer.js
@@ -1,7 +1,52 @@
 import React, { PureComponent } from 'react'
 import ProductView from './ProductView'
+import { Query } from "react-apollo";
+import ProductContentContainer from "../../components/product-content/ProductContentContainer";
 
 export class ProductViewContainer extends PureComponent {
+  constructor(props) {
+    super(props);
+
+    const { match } = this.props;
+    this.state = {
+      currentId: match.params.id,
+    };
+    this.productViewQuery = this.productViewQuery.bind(this)
+  }
+
+  productViewQuery(
+    PRODUCT_VIEW_QUERY,
+    history,
+    cartItemsHandler,
+    cartItems,
+    cartCount,
+    cartCountHandler,
+    currencySymbol,
+    quantityHandler
+  ) {
+    return (
+      <Query query={PRODUCT_VIEW_QUERY}>
+        {({ loading, data }) => {
+          if (!loading) {
+            const currentProduct = data?.product;
+            return (
+              <ProductContentContainer
+                history={history}
+                currentProduct={currentProduct}
+                currentId={this.state.currentId}
+                cartItemsHandler={cartItemsHandler}
+                cartItems={cartItems}
+                cartCount={cartCount}
+                cartCountHandler={cartCountHandler}
+                currencySymbol={currencySymbol}
+                quantityHandler={quantityHandler}
+              />
+            );
+          } else return null;
+        }}
+      </Query>
+    );
+  }
   render() {
 
     const {
@@ -36,6 +81,8 @@ export class ProductViewContainer extends PureComponent {
         changeCategory={ changeCategory }
         getTotalHandler={ getTotalHandler }
         setTotalHandler={ setTotalHandler }
+        currentId={this.state.currentId}
+        productViewQuery={this.productViewQuery}
         match={ match }
       />
     )
